pyproject.toml
[project]
name = "pp-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "mypy>=1.13.0",
    "pylint>=3.3.1",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "ruff>=0.8.0",
    "sqlalchemy-utils>=0.41.2",
    "sqlalchemy[mypy]>=2.0.36",
]
